+++ VRExpansionPlugin/Source/VRExpansionPlugin/Private/Misc/VRRenderTargetManager.cpp
@@ -80,6 +80,219 @@ void UVRRenderTargetManager::SendDrawOperations_Implementation(const TArray<FRen
 	DrawOperations();
 }
 
+
+void UVRRenderTargetManager::AddLineDrawOperation(FVector2D Point1, FVector2D Point2, FColor Color, int32 Thickness)
+{
+	FRenderManagerOperation NewOperation;
+	NewOperation.OperationType = ERenderManagerOperationType::Op_LineDraw;
+	NewOperation.Color = Color;
+	NewOperation.P1 = Point1;
+	NewOperation.P2 = Point2;
+	NewOperation.Thickness = (uint32)Thickness;
+
+	if (GetNetMode() < ENetMode::NM_Client)
+		RenderOperationStore.Add(NewOperation);
+	else
+		LocalRenderOperationStore.Add(NewOperation);
+
+	if (!DrawHandle.IsValid())
+		GetWorld()->GetTimerManager().SetTimer(DrawHandle, this, &UVRRenderTargetManager::DrawPoll, DrawRate, true);
+
+	// Send to server now
+}
+
+void UVRRenderTargetManager::AddTextureDrawOperation(FVector2D Position, UTexture2D* TextureToDisplay)
+{
+
+	if (!TextureToDisplay)
+		return;
+
+	FRenderManagerOperation NewOperation;
+	NewOperation.OperationType = ERenderManagerOperationType::Op_TexDraw;
+	NewOperation.P1 = Position;
+	NewOperation.Texture = TextureToDisplay;
+
+	if (GetNetMode() < ENetMode::NM_Client)
+		RenderOperationStore.Add(NewOperation);
+	else
+		LocalRenderOperationStore.Add(NewOperation);
+
+	if (!DrawHandle.IsValid())
+		GetWorld()->GetTimerManager().SetTimer(DrawHandle, this, &UVRRenderTargetManager::DrawPoll, DrawRate, true);
+
+	// Send to server now
+}
+
+void UVRRenderTargetManager::AddMaterialTrianglesDrawOperation(TArray<FCanvasUVTri> Tris, UMaterial* Material)
+{
+
+	if (!Tris.Num())
+		return;
+
+	FRenderManagerOperation NewOperation;
+	NewOperation.OperationType = ERenderManagerOperationType::Op_TriDraw;
+	NewOperation.Color = Tris[0].V0_Color.ToFColor(true);
+
+	NewOperation.Tris.AddUninitialized(Tris.Num());
+	int Counter = 0;
+	FRenderManagerTri RenderTri;
+	for (FCanvasUVTri Tri : Tris)
+	{
+		RenderTri.P1 = Tri.V0_Pos;
+		RenderTri.P2 = Tri.V1_Pos;
+		RenderTri.P3 = Tri.V2_Pos;
+		NewOperation.Tris[Counter++] = RenderTri;
+	}
+
+	NewOperation.Material = Material;
+
+	if (GetNetMode() < ENetMode::NM_Client)
+		RenderOperationStore.Add(NewOperation);
+	else
+		LocalRenderOperationStore.Add(NewOperation);
+
+	if (!DrawHandle.IsValid())
+		GetWorld()->GetTimerManager().SetTimer(DrawHandle, this, &UVRRenderTargetManager::DrawPoll, DrawRate, true);
+
+	// Send to server now
+}
+
+void UVRRenderTargetManager::DrawOperation(UCanvas* Canvas, const FRenderManagerOperation& Operation)
+{
+	if (LocalProxy.IsValid() && LocalProxy->OwnersID == Operation.OwnerID)
+	{
+		return;
+	}
+
+	switch (Operation.OperationType)
+	{
+	case ERenderManagerOperationType::Op_LineDraw:
+	{
+		FCanvasLineItem LineItem;
+		LineItem.Origin = FVector(Operation.P1.X, Operation.P1.Y, 0.f);
+		LineItem.EndPos = FVector(Operation.P2.X, Operation.P2.Y, 0.f);
+		LineItem.LineThickness = (float)Operation.Thickness;
+		LineItem.SetColor(Operation.Color.ReinterpretAsLinear());
+		Canvas->DrawItem(LineItem);
+	}break;
+	case ERenderManagerOperationType::Op_TexDraw:
+	{
+		if (Operation.Texture && Operation.Texture->Resource)
+		{
+			//FTexture* RenderTextureResource = (RenderBase) ? RenderBase->Resource : GWhiteTexture;
+			FCanvasTileItem TileItem(Operation.P1, Operation.Texture->Resource, FVector2D(Operation.Texture->GetSizeX(), Operation.Texture->GetSizeY()), FVector2D(0, 0), FVector2D(1.f, 1.f), ClearColor);
+			TileItem.BlendMode = FCanvas::BlendToSimpleElementBlend(EBlendMode::BLEND_Translucent);
+			Canvas->DrawItem(TileItem);
+		}
+	}break;
+	case ERenderManagerOperationType::Op_TriDraw:
+	{
+		if (Operation.Tris.Num() && Operation.Material)
+		{
+			FCanvasTriangleItem TriangleItem(FVector2D::ZeroVector, FVector2D::ZeroVector, FVector2D::ZeroVector, NULL);
+			TriangleItem.MaterialRenderProxy = Operation.Material->GetRenderProxy();
+
+			FCanvasUVTri triStore;
+			triStore.V0_Color = Operation.Color;
+			triStore.V1_Color = Operation.Color;
+			triStore.V2_Color = Operation.Color;
+
+			TriangleItem.TriangleList.Reset(Operation.Tris.Num());
+			TriangleItem.TriangleList.AddUninitialized(Operation.Tris.Num());
+			uint32 Counter = 0;
+			for (FRenderManagerTri Tri : Operation.Tris)
+			{
+				triStore.V0_Pos = Tri.P1;
+				triStore.V1_Pos = Tri.P2;
+				triStore.V2_Pos = Tri.P3;
+				TriangleItem.TriangleList[Counter++] = triStore;
+			}
+
+			Canvas->DrawItem(TriangleItem);
+		}
+	}break;
+	}
+
+}
+
+void UVRRenderTargetManager::DrawPoll()
+{
+	if (!RenderOperationStore.Num() && !LocalRenderOperationStore.Num())
+	{
+		GetWorld()->GetTimerManager().ClearTimer(DrawHandle);
+		return;
+	}
+
+	if (GetNetMode() < ENetMode::NM_Client)
+	{
+		SendDrawOperations(RenderOperationStore);
+	}
+	else
+	{
+		if (LocalRenderOperationStore.Num())
+		{
+			// Send operations to server
+			if (LocalProxy.IsValid())
+			{
+				LocalProxy->SendLocalDrawOperations(LocalRenderOperationStore);
+			}
+
+			RenderOperationStore.Append(LocalRenderOperationStore);
+			LocalRenderOperationStore.Empty();
+		}
+
+		DrawOperations();
+	}
+}
+
+void UVRRenderTargetManager::DrawOperations()
+{
+
+	if (bIsLoadingTextureBuffer)
+	{
+		if (!DrawHandle.IsValid())
+			GetWorld()->GetTimerManager().SetTimer(DrawHandle, this, &UVRRenderTargetManager::DrawPoll, DrawRate, true);
+
+		return;
+	}
+
+	UWorld* World = GetWorld();
+
+	// Reference to the Render Target resource
+	FTextureRenderTargetResource* RenderTargetResource = RenderTarget->GameThread_GetRenderTargetResource();
+
+	// Retrieve a UCanvas form the world to avoid creating a new one each time
+	UCanvas* CanvasToUse = World->GetCanvasForDrawMaterialToRenderTarget();
+
+	// Creates a new FCanvas for rendering
+	FCanvas RenderCanvas(
+		RenderTargetResource,
+		nullptr,
+		World,
+		World->FeatureLevel);
+
+	// Setup the canvas with the FCanvas reference
+	CanvasToUse->Init(RenderTarget->SizeX, RenderTarget->SizeY, nullptr, &RenderCanvas);
+	CanvasToUse->Update();
+
+	if (CanvasToUse)
+	{
+		for (const FRenderManagerOperation& opt : RenderOperationStore)
+		{
+			DrawOperation(CanvasToUse, opt);
+		}
+
+		RenderOperationStore.Empty();
+
+		// Perform the drawing
+		RenderCanvas.Flush_GameThread();
+
+		// Cleanup the FCanvas reference, to delete it
+		CanvasToUse->Canvas = NULL;
+	}
+}
+
+
 ARenderTargetReplicationProxy::ARenderTargetReplicationProxy(const FObjectInitializer& ObjectInitializer)
 	: Super(ObjectInitializer)
 {
@@ -1370,4 +1583,153 @@ bool RLE_Funcs::RLEEncodeBuffer(DataType* BufferToEncode, uint32 EncodeLength, T
 	}
 	else
 		return true;
+}
+
+template<int32 ScaleFactor, int32 MaxBitsPerComponent>
+bool WritePackedVector2D(FVector2D Value, FArchive& Ar)	// Note Value is intended to not be a reference since we are scaling it before serializing!
+{
+	check(Ar.IsSaving());
+
+	// Scale vector by quant factor first
+	Value *= ScaleFactor;
+
+	// Nan Check
+	if (Value.ContainsNaN())
+	{
+		logOrEnsureNanError(TEXT("WritePackedVector2D: Value contains NaN, clearing for safety."));
+		FVector2D	Dummy(0, 0);
+		WritePackedVector2D<ScaleFactor, MaxBitsPerComponent>(Dummy, Ar);
+		return false;
+	}
+
+	float MinV = -1073741824.0f;
+	float MaxV = 1073741760.0f;
+
+	// Some platforms have RoundToInt implementations that essentially reduces the allowed inputs to 2^31.
+	const FVector2D ClampedValue = FVector2D(FMath::Clamp(Value.X, MinV, MaxV), FMath::Clamp(Value.Y, MinV, MaxV));
+	bool bClamp = ClampedValue != Value;
+
+	// Do basically FVector::SerializeCompressed
+	int32 IntX = FMath::RoundToInt(ClampedValue.X);
+	int32 IntY = FMath::RoundToInt(ClampedValue.Y);
+
+	uint32 Bits = FMath::Clamp<uint32>(FMath::CeilLogTwo(1 + FMath::Max(FMath::Abs(IntX), FMath::Abs(IntY))), 1, MaxBitsPerComponent) - 1;
+
+	// Serialize how many bits each component will have
+	Ar.SerializeInt(Bits, MaxBitsPerComponent);
+
+	int32  Bias = 1 << (Bits + 1);
+	uint32 Max = 1 << (Bits + 2);
+	uint32 DX = IntX + Bias;
+	uint32 DY = IntY + Bias;
+
+	if (DX >= Max) { bClamp = true; DX = static_cast<int32>(DX) > 0 ? Max - 1 : 0; }
+	if (DY >= Max) { bClamp = true; DY = static_cast<int32>(DY) > 0 ? Max - 1 : 0; }
+
+	Ar.SerializeInt(DX, Max);
+	Ar.SerializeInt(DY, Max);
+
+	return !bClamp;
+}
+
+template<uint32 ScaleFactor, int32 MaxBitsPerComponent>
+bool ReadPackedVector2D(FVector2D& Value, FArchive& Ar)
+{
+	uint32 Bits = 0;
+
+	// Serialize how many bits each component will have
+	Ar.SerializeInt(Bits, MaxBitsPerComponent);
+
+	int32  Bias = 1 << (Bits + 1);
+	uint32 Max = 1 << (Bits + 2);
+	uint32 DX = 0;
+	uint32 DY = 0;
+
+	Ar.SerializeInt(DX, Max);
+	Ar.SerializeInt(DY, Max);
+
+
+	float fact = (float)ScaleFactor;
+
+	Value.X = (float)(static_cast<int32>(DX) - Bias) / fact;
+	Value.Y = (float)(static_cast<int32>(DY) - Bias) / fact;
+
+	return true;
+}
+
+bool FRenderManagerOperation::NetSerialize(FArchive& Ar, class UPackageMap* Map, bool& bOutSuccess)
+{
+	bOutSuccess = true;
+
+
+	Ar.SerializeIntPacked(OwnerID);
+	Ar.SerializeBits(&OperationType, 3);
+
+	switch (OperationType)
+	{
+	case ERenderManagerOperationType::Op_LineDraw:
+	{
+		Ar << Color;
+		Ar.SerializeIntPacked(Thickness);
+
+		if (Ar.IsSaving())
+		{
+			bOutSuccess &= WritePackedVector2D<1, 20>(P1, Ar);
+			bOutSuccess &= WritePackedVector2D<1, 20>(P2, Ar);
+		}
+		else
+		{
+			ReadPackedVector2D<1, 20>(P1, Ar);
+			ReadPackedVector2D<1, 20>(P2, Ar);
+		}
+	}break;
+	case ERenderManagerOperationType::Op_TexDraw:
+	{
+		Ar << Texture;
+
+		if (Ar.IsSaving())
+		{
+			bOutSuccess &= WritePackedVector2D<1, 20>(P1, Ar);
+		}
+		else
+		{
+			ReadPackedVector2D<1, 20>(P1, Ar);
+		}
+	}break;
+	case ERenderManagerOperationType::Op_TriDraw:
+	{
+		Ar << Color;
+		Ar << Material;
+
+		uint32 ArrayCt = Tris.Num();
+		Ar.SerializeIntPacked(ArrayCt);
+
+		if (Ar.IsLoading())
+		{
+			Tris.Reset(ArrayCt);
+			Tris.AddUninitialized(ArrayCt);
+
+			FRenderManagerTri TriTemp;
+			for (uint32 i = 0; i < ArrayCt; ++i)
+			{
+				ReadPackedVector2D<1, 20>(TriTemp.P1, Ar);
+				ReadPackedVector2D<1, 20>(TriTemp.P2, Ar);
+				ReadPackedVector2D<1, 20>(TriTemp.P3, Ar);
+				Tris[i] = TriTemp;
+			}
+		}
+		else
+		{
+			for (uint32 i = 0; i < ArrayCt; ++i)
+			{
+				WritePackedVector2D<1, 20>(Tris[i].P1, Ar);
+				WritePackedVector2D<1, 20>(Tris[i].P2, Ar);
+				WritePackedVector2D<1, 20>(Tris[i].P3, Ar);
+			}
+		}
+
+	}break;
+	}
+
+	return bOutSuccess;
 }
